backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2

sentence-transformers
faiss-cpu

transformers
torch

openai-whisper
ffmpeg-python
